requirements.txt
discord.py
aiohttp
aiosqlite
python-dotenv
playwright
requests
